# swagger.yaml
openapi: 3.0.3
info:
  title: ShortTermStayCompanyAPI
  description: API for user registration, authentication, booking, listing management, reviews, and reporting.
  version: 1.0.0
servers:
  - url: https://stsc-a3hefkewerhsfads.uaenorth-01.azurewebsites.net
    description: Azure server
paths:
...
  /insert_listing:
    post:
      summary: Insert a new listing
      tags:
        - Listing
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the listing
                numberOfPeople:
                  type: integer
                  description: Maximum number of people the listing can accommodate
                  example: 4
                country:
                  type: string
                  description: Country where the listing is located
                  example: Wonderland
                city:
                  type: string
                  description: City where the listing is located
                  example: Heartsville
                price:
                  type: number
                  format: float
                  description: Price per night for the listing
                  example: 150.0
                availableFrom:
                  type: string
                  format: date
                  description: Start date of availability
                  example: 2024-01-01
                availableTo:
                  type: string
                  format: date
                  description: End date of availability
                  example: 2024-12-31
              required:
                - numberOfPeople
                - country
                - city
                - price
                - availableFrom
                - availableTo
              example:
                title: Cozy Apartment
                numberOfPeople: 4
                country: Wonderland
                city: Heartsville
                price: 150.0
                availableFrom: 2024-01-01
                availableTo: 2024-12-31
      responses:
        '201':
          description: Listing inserted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Listing inserted successfully
        '400':
          description: Bad Request - Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example:
                      Missing required fields: numberOfPeople, country
        '401':
          description: Unauthorized - Missing or invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Missing or invalid token
        '403':
          description: Forbidden - User does not have host role
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Access forbidden insufficient permissions
  /listing/listings:
    get:
      summary: Retrieve a paginated list of listings
      tags:
        - Listing
      parameters:
        - name: page
          in: query
          description: "Page number (default: 1)"
          required: false
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          description: "Listings per page (default: 10, max: 100)"
          required: false
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Listings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetListingsResponse'
        '400':
          description: Bad Request - Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPage:
                  summary: Invalid page number
                  value:
                    message: Page number must be 1 or greater.
                invalidPerPage:
                  summary: Invalid per_page value
                  value:
                    message: per_page must be between 1 and 100.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GetListingsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
        meta:
          type: object
          properties:
            page:
              type: integer
              example: 1
            per_page:
              type: integer
              example: 10
            total_pages:
              type: integer
              example: 5
            total_items:
              type: integer
              example: 50
            has_next:
              type: boolean
              example: true
            has_prev:
              type: boolean
              example: false
            next_page:
              type: integer
              example: 2
            prev_page:
              type: integer
              nullable: true
              example: null
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Error message detailing what went wrong
    Listing:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: Cozy Apartment
        numberOfPeople:
          type: integer
          minimum: 1
          maximum: 32
          example: 4
        country:
          type: string
          example: Wonderland
        city:
          type: string
          example: Heartsville
        price:
          type: number
          format: float
          example: 150.0